# LU decomposition: random numbers initialisation in parallel