This patch adds parallelisation to random numbers initialisation in LU decomposition benchmark. Performance comparison before and after the patch is below:

| Cores | Time before (s) | Speedup      | Time after (s) | Speedup     |
|-------|-----------------|--------------|----------------|-------------|
| 1     | 35.86           | 0.9428332404 | 33.81          | 1           |
| 2     | 18.31           | 1.84653195   | 17.13          | 1.973730298 |
| 4     | 9.48            | 3.566455696  | 8.74           | 3.868421053 |
| 8     | 5.1             | 6.629411765  | 4.64           | 7.286637931 |
| 12    | 3.9             | 8.669230769  | 3.5            | 9.66        |
| 16    | 3.11            | 10.87138264  | 2.74           | 12.33941606 |
| 20    | 2.68            | 12.61567164  | 2.32           | 14.57327586 |
| 24    | 2.58            | 13.10465116  | 2.21           | 15.29864253 |